refactor(autogroups): Extract functions to groups duck
It's easier to test them in isolation
